chore: move worker listeners closer to the worker
